Fix config init + install spec device
Was missing the install spec target to be synced with the
cfg.instlal.target to get the proper target when the device is not
coming from the cloud-config.

So basically if your target came from the cloudconfig it would be ok on
unmarshall the spec would get filled, but on manual install that value
is missing from the cloud config files and set on the config manually so
we need to sync that up once we generate a new install spec.

Also mark some fields on the Config structure as non marshalled,
otherwise things like the logger or the fs would get zeroed if those
values are not present on the cloud config.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>